# apps/backend/openapi.test3.yaml
openapi: 3.0.3
info:
  title: X-Ear CRM API (auto-generated)
  version: 0.0.0-auto
servers:
- url: http://localhost:5003
paths:
...
  /api/inventory:
    get:
      summary: Get all inventory items with optional filtering
      operationId: inventory_get_all_inventory
      tags:
      - Inventory
      responses:
        '200':
          description: Success
    post:
      summary: Create a new inventory item
      operationId: inventory_create_inventory_item
      tags:
      - Inventory
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/inventory/{item_id}:
    get:
      summary: Get a single inventory item by ID
      operationId: inventory_get_inventory_item
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
    put:
      summary: Update an existing inventory item
      operationId: inventory_update_inventory_item
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
    delete:
      summary: Delete an inventory item
      operationId: inventory_delete_inventory_item
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Success
  /api/inventory/{item_id}/activity:
    post:
      summary: Log an activity for an inventory item
      operationId: inventory_log_inventory_activity
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
    get:
      summary: Get activity log for an inventory item
      operationId: inventory_get_inventory_activities
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
  /api/inventory/{item_id}/assign:
    post:
      summary: Assign inventory item to a patient (reduces stock)
      operationId: inventory_assign_to_patient
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
  /api/inventory/{item_id}/serials:
    post:
      summary: Add serial numbers to an inventory item
      operationId: inventory_add_serial_numbers
      tags:
      - Inventory
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
...